Dojo Description
Hi,

Just wanted to comment  that the Dojo description is unclear.  A building like the Runestone provides every city with a Runestone a bonus when a tile is pillaged; however, the Dojo only provides a bonus to the city that produced the unit that is getting promoted.  This should be made clear in the building's description.  That's all - thanks!